Updating with scatters clickdata
I just added the clickData event for the scatterplot, but I realized that when you click on the scatterplot, the clickData for the choropleth won't work because of the if else statements, I have to figure out how to overcome this, because there's no way to clear theclickData. In addition, there is a tab for the rankpath, which I'm doing.