Creating Web-STARS-UI.py
This is a version to deploy the current dashboard made in Dash with a simple layout. Su, I chose to supress the reactives of the scatterPlot (clickdata) and let only the ones with the choropleth, because since there's no way (at least, not simple way) to clear the clickData or select data, there's no way to keep the user changing to click on either choropleth or scatter plot freely, because the user will eventually not able to refresh using the other option.